1. Don't download if files already exist
2. add a flag to the parse_and_download function
   "keep_link_string_as_name" -- if false keep link basename as file name
	else keep the link string as filename(default is true)

+++ Python Scripts/Download all pdfs/download-all-pdfs.py
@@ -12,18 +12,21 @@ except ImportError:
     sys.exit(0)
 
 
-def parse_and_download(url, download_path):
+def parse_and_download(url, download_path, keep_link_string_as_name=True):
     '''
     open, parse a url and try to download all pdfs on that page
 
     Arguments:
         url -- url from where you want to download pdfs
         download_path -- location where to keep downloaded files
+
+    Keyword-arguemnts:
+        keep_link_string_as_name -- if false keep link basename as file name
+        else keep the link string as filename(default is true)
     '''
 
     # open and read the url
     html = urllib2.urlopen(url).read()
-
     # creating a BeautifulSoup object, which represents
     # the document as a nested data structure:
     soup = BeautifulSoup(html, "lxml")
@@ -41,24 +44,32 @@ def parse_and_download(url, download_path):
         extension = os.path.splitext(basename)[1]
 
         if extension == ".pdf":
-            filepath = os.path.join(download_path, basename)
 
-            file = open(filepath, "wb")  # open filestream
-            try:
-                print "\n%d: Downloading %s" % (num_files, basename)
-                # open a connection to this pdf
-                pdf = urllib2.urlopen(link)
-                file.write(pdf.read())
-                num_files += 1
+            if tag.string is not None and keep_link_string_as_name is True:
+                filepath = os.path.join(download_path, tag.string+".pdf")
+            else:
+                filepath = os.path.join(download_path, basename)
 
-            except Exception as e:
-                failed += 1
-                print "\nCouldn't download %s" % (basename)
-                print e
-                os.remove(filepath)
+            if not os.path.exists(filepath):
+                file = open(filepath, "wb")  # open filestream
+                try:
+                    print "\n%d: Downloading %s" % (num_files, basename)
+                    # open a connection to this pdf
+                    pdf = urllib2.urlopen(link)
+                    file.write(pdf.read())
+                    num_files += 1
 
-            finally:
-                file.close()  # close file stream
+                except Exception as e:
+                    failed += 1
+                    print "\nCouldn't download %s" % (basename)
+                    print e
+                    os.remove(filepath)
+
+                finally:
+                    file.close()  # close file stream
+            else:
+                print "\nSkipping %s" % (basename)
+                print "File aleady exists"
 
     print "\n\n Successfully downloaded %d files" % (num_files - 1)
     print "\n %d files failed to download\n" % failed
@@ -81,7 +92,7 @@ if __name__ == '__main__':
         url = inputs[1].strip()
         download_path = inputs[2].strip()
         try:
-            parse_and_download(url, download_path)
+            parse_and_download(url, download_path, keep_link_string_as_name=True)
 
         except KeyboardInterrupt:
             print "\nExiting..."
